commit 2202

+++ src/time-options/time-options.service.ts
@@ -1,5 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { Prisma } from '@prisma/client';
 import { PrismaService } from '../prisma/prisma.service';
+import { CreateTimeOptionDto } from './dto/create-time-options.dto';
 
 @Injectable()
 export class TimeOptionsService {
@@ -16,33 +18,45 @@ export class TimeOptionsService {
     });
   }
 
-  async create({ day, time }: { day: number; time: Date }) {
-    day = Number(day);
-    time = new Date(time);
+  async create(data: CreateTimeOptionDto) {
+    const { day, time } = data;
 
     return this.prisma.timeOption.create({
       data: {
-        day,
-        time,
+        day: day,
+        time: time,
       },
     });
   }
 
   async update(id, { day, time }: { day: number; time: Date }) {
-    day = Number(day);
-    time = new Date(time);
+    const dataTimeOption = {} as Prisma.TimeOptionUpdateInput;
+
+    if (day) {
+      dataTimeOption.day = +day;
+    }
+
+    if (time) {
+      dataTimeOption.time = new Date(time);
+    }
+
     return await this.prisma.timeOption.update({
       where: {
         id,
       },
-      data: {
-        day,
-        time,
-      },
+      data: dataTimeOption,
     });
   }
 
   async delete(id) {
+    if (isNaN(id)) {
+      throw new BadRequestException('Id is required');
+    }
+    const timeOption = await this.getById(id);
+
+    if (!timeOption) {
+      throw new BadRequestException('Time option not found');
+    }
     return await this.prisma.timeOption.delete({
       where: {
         id,
